requirements.txt
spacy~=3.7.0
asent~=0.8.0
Django~=4.2.5
djangorestframework~=3.14.0